life-functions-app/requirements.txt
azure-functions
numpy
requests